Moving common code for TestCreateTable and TestTruncateTable TestCreateTable and TestTruncateTable are mostly the same for HBase 1 and HBase 2.  Moving comon code into Abstract* tests in bigtable-hbase-integration-tests-common, and extending in hbase1 and hbase2 test projects.
